+++ src/test/java/leetCode/matrix/Shift2DGridTest.java
@@ -1,4 +1,4 @@
-package leetCode.array;
+package leetCode.matrix;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
